\docType{data}
\name{acc2geneMTB}
\alias{acc2geneMTB}
\title{miRNA targets from miRTarBase}
\description{
  Lists containing the target genes or transcripts of each
  miRNAs.
}
\details{
  Data were downloaded form miRTarBase
  http://mirtarbase.mbc.nctu.edu.tw/
}
\examples{
acc2geneMTB[1:3]
}
\author{
  David Montaner \email{dmontaner@cipf.es}
}
\seealso{
  \code{\link{miRTarBaseDownloadDate}}
}
\keyword{accession}
\keyword{convert}
\keyword{datasets}
\keyword{gene}
\keyword{id}
\keyword{microRNA}
\keyword{miRNA}
\keyword{target}
\keyword{transcript}

