\docType{data}
\name{miRBaseDownloadDate}
\alias{miRBaseDownloadDate}
\title{miRBase download date}
\description{
  A variable keeping the date when the miRBase was
  downloaded.
}
\examples{
miRBaseDownloadDate
}
\author{
  David Montaner \email{dmontaner@cipf.es}.
}
\keyword{datasets}
\keyword{date}
\keyword{download}

