\name{mirbVersions}
\alias{mirbVersions}
\title{Show all available miRBase versions}
\usage{
  mirbVersions()
}
\value{
  A character vector of available versions.
}
\description{
  A function to display all available miRBase versions.
}
\examples{
mirbVersions ()
}
\author{
  David Montaner <\email{dmontaner@cipf.es}>
}
\seealso{
  \code{\link{mirIDmat}} \code{\link{id2mirLast}}
}

