\docType{data}
\name{targetScanDownloadDate}
\alias{targetScanDownloadDate}
\title{TargetScan download date}
\description{
  A variable keeping the date when the TargetScan was
  downloaded.
}
\examples{
targetScanDownloadDate
}
\author{
  David Montaner \email{dmontaner@cipf.es}.
}
\seealso{
  miRBaseDownloadDate
}
\keyword{datasets}
\keyword{date}
\keyword{download}

